Pull out the centroid writer code
